+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/sidebar-layout.tsx
@@ -24,8 +24,6 @@ import {
 } from "@stackframe/stack-ui";
 import {
   Book,
-  Box,
-  CreditCard,
   Globe,
   KeyRound,
   Link as LinkIcon,
@@ -40,7 +38,7 @@ import {
   SquarePen,
   User,
   Users,
-  Webhook,
+  Webhook
 } from "lucide-react";
 import { useTheme } from "next-themes";
 import { usePathname } from "next/navigation";
@@ -214,10 +212,10 @@ const navigationItems: (Label | Item | Hidden)[] = [
     regex: /^\/projects\/[^\/]+\/email-themes\/[^\/]+$/,
     type: 'hidden',
   },
+  /*
   {
     name: "Payments",
     type: 'label',
-    requiresDevFeatureFlag: true,
   },
   {
     name: "Payments",
@@ -225,24 +223,7 @@ const navigationItems: (Label | Item | Hidden)[] = [
     regex: /^\/projects\/[^\/]+\/payments$/,
     icon: CreditCard,
     type: 'item',
-    requiresDevFeatureFlag: true,
-  },
-  {
-    name: "Offers",
-    href: "/payments/offers",
-    regex: /^\/projects\/[^\/]+\/payments\/offers$/,
-    icon: SquarePen,
-    type: 'item',
-    requiresDevFeatureFlag: true,
-  },
-  {
-    name: "Items",
-    href: "/payments/items",
-    regex: /^\/projects\/[^\/]+\/payments\/items$/,
-    icon: Box,
-    type: 'item',
-    requiresDevFeatureFlag: true,
-  },
+  },*/
   {
     name: "Configuration",
     type: 'label'
@@ -571,7 +552,7 @@ export default function SidebarLayout(props: { projectId: string, children?: Rea
         </div>
 
         {/* Content Body - Normal scrolling */}
-        <div className="flex-grow relative">
+        <div className="flex-grow relative flex flex-col">
           {props.children}
         </div>
       </div>