+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @stackframe/stack-backend
 
+## 2.8.37
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.37
+
+## 2.8.36
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.36
+
+## 2.8.35
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.35
+
 ## 2.8.34
 
 ### Patch Changes